serial: sh-sci: Require a device per port mapping.
In the olden days it was possible to register a map of ports per device,
but this has long since been abandoned due to the desire to match up with
clkdev and other functionality. As a result, all of the in-tree users
have for some time already been migrated off of such behaviour, while
we've left support code in place to deal with either case. Dropping the
code permits for quite a bit of simplification, so we do that now before
any users of the old interface are able to be added back in.

Signed-off-by: Paul Mundt <lethal@linux-sh.org>
(cherry picked from commit d535a2305facf9b49a65bd412c6deaaf4d4316f2)

Conflicts:

	drivers/serial/sh-sci.c

Signed-off-by: Nguyen Hong Ky <bimtocngoan@gmail.com>