serial: sh-sci: Kill off the special earlyprintk device.
This reworks some of the code a bit so that we always wrap in to the
global port array instead of flipping between it and the special early
device. This will make it easier to split the initialization in to early
and late parts where we can simply initialize the remaining bits of the
partially-initialized port coming off of the early console later in the
boot process.

Signed-off-by: Paul Mundt <lethal@linux-sh.org>
(cherry picked from commit 906b17dc089f7fa87e37a9cfe6ee185efc90e0da)

Signed-off-by: Nguyen Hong Ky <bimtocngoan@gmail.com>