sh: sh-sci: Fix double initialization by serial_console_setup
The driver is initialized in a state with an unknown value by
serial_console_setup. And initialization fails.

This is caused by the initialization by sci_console_init.
This function does not seem to be necessary for the present sh-sci driver.

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu.yj@renesas.com>
Signed-off-by: Paul Mundt <lethal@linux-sh.org>
(cherry picked from commit 6a8c979935f1955707fe79909cb3baf9575282f4)

Signed-off-by: Nguyen Hong Ky <bimtocngoan@gmail.com>