PM / Runtime: Update documentation of interactions with system sleep
The documents describing the interactions between runtime PM and
system sleep generally refer to the model in which the system sleep
state is entered through a global firmware or hardware operation.
As a result, some recommendations given in there are not entirely
suitable for systems in which this is not the case.  Update the
documentation to take the existence of those systems into account.

Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>
Reviewed-by: Kevin Hilman <khilman@ti.com>
(cherry picked from commit 455716e9b12ba93e93181ac88bef62e4eb5ac66c)

Signed-off-by: Simon Horman <horms@verge.net.au>